Validacion de email y redondeo de decimales

+++ Source/Administrador/database/migrations/08_05_2016_000000_crear_tabla_productos.php
@@ -24,7 +24,7 @@ class CrearTablaProductos extends Migration
 				$table->integer('marca');
 				$table->integer('estado');
 				$table->integer('categoria');
-				$table->integer('precio');
+				$table->double('precio',10,2);
         });
     }
